flytt interface til egen fil
